Rename //rules_python package to //packaging
This avoids a name clash between the package name and the repo name, which
prevented whl_test from running in the Bazel federation's CI.

Regenerated the par files.

Tested that this still works in environments where the system Python has a
library named "packaging" installed, by creating a virtualenv, pip installing
"packaging", and running `bazel test //examples/...`.

Fixes #227.
