+++ update_tools.sh
@@ -25,9 +25,9 @@ if [ "$#" -eq 0 ] ; then
   docker build --no-cache -f tools/update_tools/Dockerfile --tag rules_python:update_tools .
   docker run -v"$PWD":/opt/rules_python_source rules_python:update_tools
 elif [ "$#" -eq 1 -a "$1" == "--nodocker" ] ; then
-  bazel build //rules_python:piptool.par //rules_python:whltool.par
-  cp bazel-bin/rules_python/piptool.par tools/piptool.par
-  cp bazel-bin/rules_python/whltool.par tools/whltool.par
+  bazel build //packaging:piptool.par //packaging:whltool.par
+  cp bazel-bin/packaging/piptool.par tools/piptool.par
+  cp bazel-bin/packaging/whltool.par tools/whltool.par
 else
   usage
 fi
